Edit product media dimensions
In this case to make the functionality of this snippet simple, we asume
that the application will force the product owner to choose images that
have similar dimensions (width should be equal to height) in order for
zoom to work.
Later we'll add support to any kind of image (e.g. landscape)